Bound Abyss shifts
Convert abyss depth into an int, rip up all the phase logic (sorry).

Each procedural generator declares a changepoint, a lower bound estimate
on where the terrain type can change. This way, we can partially
recompute the abyss without experiencing aliasing.

+++ crawl-ref/source/abyss.h
@@ -17,9 +17,10 @@ struct abyss_state
 {
     coord_def major_coord;
     uint32_t seed;
-    double phase;
-    double depth;
+    uint32_t phase;
+    uint32_t depth;
 };
+
 void abyss_morph(double duration);
 void push_features_to_abyss();
